Refactor: Add TaskList class
